requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.2

# OpenAI and LangChain
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langchain-core==0.1.3
langchain-text-splitters==0.0.1

# Vector store and embeddings
chromadb==0.5.18
tokenizers==0.13.3
sentence-transformers==2.2.2

# Document processing
pypdf==3.17.1
python-docx==1.0.1
Pillow==10.1.0
PyMuPDF==1.23.8  # Also known as fitz
pytesseract==0.3.10  # For OCR capabilities
pdf2image==1.16.3   # For PDF image extraction

# File and media handling
python-magic==0.4.27; platform_system != "Windows"  # For Unix systems
python-magic-bin==0.4.14; platform_system == "Windows"  # For Windows
moviepy==1.0.3  # For video processing
pydub==0.25.1   # For audio processing
antiword==0.3.6; platform_system == "Linux"
# Database and async
sqlalchemy==2.0.23
aiosqlite==0.19.0
aiofiles==23.2.1
asyncio==3.4.3
aiohttp==3.9.1

# Data processing and analysis
numpy==1.26.2
pandas==2.1.3
scipy==1.11.4

# Configuration and parsing
PyYAML==6.0.1
configparser==6.0.0
typing-extensions==4.8.0

# CLI and interface
rich==13.7.0
typer==0.9.0
click==8.1.7
tqdm==4.66.1

# Utilities
tenacity==8.2.3  # For retrying operations
python-dateutil==2.8.2
requests==2.31.0
urllib3==2.1.0
certifi==2023.11.17

# Machine Learning and Optimization
scikit-learn==1.3.2
torch==2.5.1     # Required for some embeddings
transformers==4.35.2  # Required for some LLM operations

# Testing and Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
mypy==1.7.1
black==23.11.0
isort==5.12.0

# Type checking support
types-PyYAML==6.0.12.12
types-python-dateutil==2.8.19.14
types-requests==2.31.0.10
types-setuptools==69.0.0.0
types-urllib3==1.26.25.14

# Optional but recommended for performance
onnxruntime==1.16.3; platform_system == "Windows"
onnxruntime-gpu==1.16.3; platform_system != "Windows"  # For systems with GPU

# Security
cryptography==41.0.5
python-jose[cryptography]==3.3.0
passlib==1.7.4

# Environment and path management
pathlib==1.0.1
environs==9.5.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0